mobile map legend
