+++ packages/tupaia-web/src/features/Map/MapLegend/MobileMapLegend.tsx
@@ -10,8 +10,10 @@ import { MOBILE_BREAKPOINT } from '../../../constants';
 
 const Wrapper = styled.div`
   position: absolute;
-  bottom: 1rem;
+  pointer-events: auto;
+  bottom: 0.8rem;
   right: 1rem;
+  padding: 0;
   @media screen and (min-width: ${MOBILE_BREAKPOINT}) {
     display: none;
   }
@@ -30,10 +32,13 @@ const ExpandIcon = styled(ExpandLess)`
 `;
 
 const ExpandedLegend = styled.div`
+  display: block;
   background-color: ${({ theme }) => theme.mobile.background};
-  height: 20rem;
-  width: 12rem;
+  min-height: 10rem;
+  min-width: 10rem;
   border-radius: 0.5rem;
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem;
 `;
 
 const CloseButton = styled(IconButton).attrs({
@@ -44,7 +49,7 @@ const CloseButton = styled(IconButton).attrs({
   right: 0;
 `;
 
-export const MobileMapLegend = () => {
+export const MobileMapLegend = ({ Legend }) => {
   const [expanded, setExpanded] = useState(false);
   const toggleExpanded = () => {
     setExpanded(!expanded);
@@ -56,6 +61,7 @@ export const MobileMapLegend = () => {
           <CloseButton onClick={toggleExpanded} aria-label="Close legend">
             <Close />
           </CloseButton>
+          {Legend}
         </ExpandedLegend>
       ) : (
         <MapLegendExpandButton onClick={toggleExpanded}>
